Move transcription segment
