+++ lib/src/types/transcription_segment.dart
@@ -1,0 +1,34 @@
+// Copyright 2025 LiveKit, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import 'package:flutter/foundation.dart';
+
+@immutable
+class TranscriptionSegment {
+  final String id;
+  final String text;
+  final DateTime firstReceivedTime;
+  final DateTime lastReceivedTime;
+  final bool isFinal;
+  final String language;
+
+  const TranscriptionSegment({
+    required this.id,
+    required this.text,
+    required this.firstReceivedTime,
+    required this.lastReceivedTime,
+    required this.isFinal,
+    required this.language,
+  });
+}
